prop-settings-3: refactor schema into own module, improve output of nc-validate
